Fix data type of GC_ExtensionsBase::usablePhysicalMemory
GC_ExtensionsBase::usablePhysicalMemory should be of same type as the value
returned by omrsysinfo_get_addressable_physical_memory() which is uint64_t.

+++ gc/base/GCExtensionsBase.hpp
@@ -546,7 +546,7 @@ public:
 
 	uintptr_t overflowSafeAllocSize;
 
-	uintptr_t usablePhysicalMemory; /**< Physical memory available to the process */
+	uint64_t usablePhysicalMemory; /**< Physical memory available to the process */
 
 #if defined(OMR_GC_REALTIME)
 	/* Parameters */
